+++ packages/sdk/src/get_sdk.ts
@@ -1,11 +1,13 @@
-import Sdk from './sdk';
-import airtableInterface from './injected/airtable_interface';
+import SdkType from './sdk';
 
 // TODO(alex): prevent sdk sharing across invocations of the same lambda container
-let sdk: Sdk;
+let sdk: SdkType;
+
 /** @hidden */
-export default function getSdk(): Sdk {
+export default function getSdk(): SdkType {
     if (!sdk) {
+        const airtableInterface = require('./injected/airtable_interface').default;
+        const Sdk = require('./sdk').default;
         sdk = new Sdk(airtableInterface);
     }