+++ packages/sdk/src/stats/block_stats.ts
@@ -1,21 +1,15 @@
-import airtableInterface, {AirtableInterface} from '../injected/airtable_interface';
+import getSdk from '../get_sdk';
 
 // NOTE(richsinn): this client is effectively an implementation of the StatsClient interface
 // for Datadog. For example, see https://github.com/Hyperbase/hyperbase/blob/development/client_server_shared/stats/stats_client.tsx
 
 /** @hidden */
 class BlockStats {
-    private _airtableInterface: AirtableInterface;
-
-    constructor(injectedAirtableInterface: AirtableInterface) {
-        this._airtableInterface = injectedAirtableInterface;
-    }
-
     increment(stat: string, tags?: {[key: string]: string}) {
         this.incrementBy(stat, 1, tags);
     }
     incrementBy(stat: string, value: number, tags?: {[key: string]: string}) {
-        this._airtableInterface.sendStat({
+        getSdk().__airtableInterface.sendStat({
             metricType: 'incrementBy',
             stat,
             value,
@@ -26,7 +20,7 @@ class BlockStats {
         this.decrementBy(stat, 1, tags);
     }
     decrementBy(stat: string, value: number, tags?: {[key: string]: string}) {
-        this._airtableInterface.sendStat({
+        getSdk().__airtableInterface.sendStat({
             metricType: 'decrementBy',
             stat,
             value,
@@ -34,7 +28,7 @@ class BlockStats {
         });
     }
     gauge(stat: string, value: number, tags?: {[key: string]: string}) {
-        this._airtableInterface.sendStat({
+        getSdk().__airtableInterface.sendStat({
             metricType: 'gauge',
             stat,
             value,
@@ -42,7 +36,7 @@ class BlockStats {
         });
     }
     histogram(stat: string, value: number, tags?: {[key: string]: string}) {
-        this._airtableInterface.sendStat({
+        getSdk().__airtableInterface.sendStat({
             metricType: 'histogram',
             stat,
             value,
@@ -58,7 +52,7 @@ class BlockStats {
         });
     }
     timing(stat: string, time: number, tags?: {[key: string]: string}) {
-        this._airtableInterface.sendStat({
+        getSdk().__airtableInterface.sendStat({
             metricType: 'timing',
             stat,
             value: time,
@@ -71,4 +65,4 @@ class BlockStats {
     }
 }
 
-export const blockStats = new BlockStats(airtableInterface);
+export const blockStats = new BlockStats();