+++ packages/sdk/test/ui/icon.test.tsx
@@ -1,0 +1,9 @@
+import React from 'react';
+import {mount} from 'enzyme';
+import {Icon} from '../../src/ui/ui';
+
+describe('Icon', () => {
+    it('renders outside of a blocks context', () => {
+        mount(<Icon name="bell" />);
+    });
+});