+++ packages/sdk/test/ui/ui.test.tsx
@@ -1,0 +1,5 @@
+describe('ui', () => {
+    it('can be imported outside of a blocks context', () => {
+        require('../../src/ui/ui');
+    });
+});