+++ packages/sdk/test/unstable_private_utils.test.ts
@@ -1,0 +1,5 @@
+describe('unstable_private_utils', () => {
+    it('can be imported outside of a blocks context', () => {
+        require('../src/unstable_private_utils');
+    });
+});