+++ packages/sdk/unstable_private_utils.d.ts
@@ -1,4 +1,1 @@
-export * from './dist/types/src/private_utils';
-export * from './dist/types/src/error_utils';
-export * from './dist/types/src/event_tracker';
-export * from './dist/types/src/stats/block_stats';
+export * from './dist/types/src/unstable_private_utils';