+++ packages/sdk/unstable_private_utils.js
@@ -1,7 +1,1 @@
-module.exports = Object.assign(
-    {},
-    require('./dist/cjs/private_utils'),
-    require('./dist/cjs/error_utils'),
-    require('./dist/cjs/event_tracker'),
-    require('./dist/cjs/stats/block_stats'),
-);
+module.exports = require('./dist/cjs/unstable_private_utils');